experimental-plugins/agentic-core/pyproject.toml
[project]
name = "agentic-core"
version = "0.1.0"
description = "Foundational framework for AI agent orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "ruamel.yaml>=0.18.0",
    "jinja2>=3.1.0",
    "asyncpg>=0.29.0",
    "confluent-kafka>=2.3.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
memory = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.26.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
agentic = "agentic_core.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_core"]

[tool.hatch.build.targets.wheel.shared-data]
"docker" = "share/agentic-core/docker"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
